+++ Changelog.md
@@ -17,6 +17,8 @@ Enhancements:
 * Mocks expectations can now set a custom failure message.
   (Benoit Tigeot and Nicolas Zermati, #1156)
 * `aggregate_failures` now shows the backtrace line for each failure. (Fabricio Bedin, #1163)
+* Support multiple combinations of `yield_control` modifiers like `at_least`, `at_most`.
+  (Jon Rowe, #1169)
 
 ### 3.9.0 / 2019-10-08
 [Full Changelog](http://github.com/rspec/rspec-expectations/compare/v3.8.6...v3.9.0)